third_party/blink/web_tests/http/tests/devtools/sources/debugger/dynamic-script-tag-expected.txt

Tests that inline scripts and document.write scripts get different uiSourceCodes with different URLs.

Running: testOpenDevToolsAfterLoad
Both dynamic scripts were loaded.
Number of non-inline scripts with document url as a sourceURL: 0.

Running: testOpenDevToolsThenReload
Reloading page.
The first script with document url as a sourceURL to be seen is inline script.
Page reloaded.
